refactored MainActivity

+++ app/src/main/java/ru/neosvet/utils/Const.java
@@ -6,7 +6,7 @@ package ru.neosvet.utils;
 
 public interface Const {
     int MAX_ON_PAGE = 15, TIMEOUT = 20;
-    byte TURN_OFF = -1, SCREEN_MENU = 0, SCREEN_CALENDAR = 1, SCREEN_SUMMARY = 2;
+    int TURN_OFF = -1, SCREEN_MENU = 0, SCREEN_CALENDAR = 1, SCREEN_SUMMARY = 2;
     String mailto = "mailto:neosvet333@gmail.com?subject=Приложение «Весть Века»&body=";
     String LIGHT = "/style/light.css", DARK = "/style/dark.css", ENCODING = "cp1251";
     String COMMA = ",", N = "\n", NN = "\n\n", AND = "&", SEARCH = "search";
